chore(repo): change core proj to urlboxSDK to match namespace
