+++ Urlbox.MsTest/Urlbox.MsTest.csproj
@@ -20,6 +20,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="../Urlbox/Urlbox/Urlbox.csproj" />
+    <ProjectReference Include="../Urlbox/UrlboxSDK/Urlbox.csproj" />
   </ItemGroup>
 </Project>
