+++ Urlbox/UrlboxSDK/Metadata/Resource/UrlboxMetadata.cs
