+++ Urlbox/UrlboxSDK/Resource/IUrlbox.cs
