+++ Urlbox/UrlboxSDK/Resource/UrlboxResponse.cs
